Lambda code for setting container instance state
Lambda code is triggered by the ASG notification event of instance termination. It obtains the container instance ID, and sets the state of the container instance to DRAINING. This will ensure that the Service Scheduler will be able to vacate the tasks from the instance that is DRAINING and find an alternate ECS container instance to place tasks.